Fix python crash
`m` in `export_geometry` was still referred to via references to contained
Vertex instances but is deleted when going out of scope and will therefore
not be a valid reference during the DdlDocumet.write call.

By keeping a reference in the corresponding Mesh DdlStructure subclass, we
guarantee that the mesh data will still be available during write of the
file.

Signed-off-by: Squareys <Squareys@googlemail.com>